refactor connector create_pg_client logic
