+++ src/connector/src/connector_common/mod.rs
@@ -25,4 +25,8 @@ pub use common::{
 };
 
 mod iceberg;
+mod maybe_tls_connector;
+mod postgres;
+
 pub use iceberg::IcebergCommon;
+pub use postgres::{create_pg_client, SslMode};
